α Fixed a horrible runtime issue that occurred during GUI integration
The problem was related to the fact that it was necessary to create Texture2D in an anonymous namespace, but this is actually a global or static initialization area. Global or static variables are initialized before the execution of main, so before the OpenGL context is configured. However, in the Texture2D constructor, glGenTextures was called, which was nullptr and therefore caused indecipherable errors at runtime due to the fact that opengl32, the dll needed on windows is built without symbols, that is, it was compiled without including debug information or debug symbols useful for code tracing during runtime.